pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-unicz"
version = "0.1.0"
description = "RAG application for Unicz"
requires-python = ">=3.11"
dependencies = [
    "haystack-ai>=2.6",
    "qdrant-client",
    "fastembed",
    "ollama",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.7",
    "pydantic-settings",
    "pymupdf",
    "python-docx",
    "pandas",
    "PyYAML",
    "httpx",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py311"
